number-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-multipart>=0.0.17
ultralytics>=8.3.0
opencv-python>=4.10.0
numpy>=1.26.0
pillow>=10.0.0
torch
torchvision
scikit-learn>=1.3.0
tensorflow>=2.15.0
